pyproject.toml
[tool.poetry]
name = "portfolio_generator"
version = "0.2.0"
description = ""
authors = ["atu4403  <73111778+atu4403@users.noreply.github.com>"]
homepage = "https://github.com/atu4403"
repository = "https://github.com/atu4403/portfolio_generator"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.3"
Jinja2 = "^3.0.3"
requests = "^2.26.0"
PyYAML = "^6.0"
arrow = "^1.2.1"
adash = "^1.2.0"
xmltodict = "^0.12.0"
feedparser = "^6.0.8"
Cerberus = "^1.3.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
exclude = '''
(
    migrations
    | .mypy_cache
    | .pytest_cache
    | .tox
    | venv
)
'''

[flake8]
max-line-length = 99
